move EVM contract to flow integration
