+++ languageserver/server/server.go
@@ -184,6 +184,8 @@ type Server struct {
 	// initializationOptionsHandlers are the functions that are used to handle initialization options sent by the client
 	initializationOptionsHandlers []InitializationOptionsHandler
 	accessCheckMode               sema.AccessCheckMode
+	// extendedStandardLibraryValues are the standard library values that are provided in addition to the default standard library
+	extendedStandardLibraryValues []stdlib.StandardLibraryValue
 	// reportCrashes decides when the crash is detected should it be reported
 	reportCrashes bool
 	// checkerStandardConfig is a config used to check contracts and transactions
@@ -276,6 +278,18 @@ func WithMemberAccountAccessHandler(handler sema.MemberAccountAccessHandlerFunc)
 	return func(server *Server) error {
 		server.checkerStandardConfig.MemberAccountAccessHandler = handler
 		server.checkerScriptConfig.MemberAccountAccessHandler = handler
+		return nil
+	}
+}
+
+// WithStandardLibraryValues returns a server option that adds the given function
+// as a function that is used to resolve standard library values (in addition to the default standard library)
+//
+// This is used to provide parts of the standard library that are present in an environment,
+// but are not native to the language.
+func WithExtendedStandardLibraryValues(values ...stdlib.StandardLibraryValue) Option {
+	return func(server *Server) error {
+		server.extendedStandardLibraryValues = append(server.extendedStandardLibraryValues, values...)
 		return nil
 	}
 }
